PythonInstance/requirements.txt
py-cord==2.0.0b7
python-decouple==3.6
discord-together